Touchless DQ/requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
pandas
openpyxl
